requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
oauthlib==3.2.2
passlib==1.7.4
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.7.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
svgwrite==1.4.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
